--- kzg/src/finite.rs
@@ -1,9 +1,0 @@
-use super::Fr;
-
-pub fn add_fr(first: Fr, second: Fr) -> Fr {
-    let mut sum = Fr::default();
-    unsafe {
-        blst::blst_fr_add(&mut sum, &first, &second);
-    }
-    sum
-}
