--- kzg10-rust/Cargo.toml
@@ -1,39 +1,0 @@
-[package]
-name = "mcl_rust"
-version = "0.0.1"
-authors = ["MITSUNARI Shigeo <herumi@nifty.com>"]
-description = "a wrapper class/function of a pairing library; https://github.com/herumi/mcl"
-license = "BSD-3-Clause OR MIT OR Apache-2.0"
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dev-dependencies]
-criterion = "0.3"
-rand = "0.5"
-
-[[bench]]
-name = "bench_fft"
-harness = false
-
-[[bench]]
-name = "bench_proof_check"
-harness = false
-
-[[bench]]
-name = "bench_das"
-harness = false
-
-[[bench]]
-name = "bench_proof_gen"
-harness = false
-
-[[bench]]
-name = "bench_commits"
-harness = false
-
-[dependencies]
-cc = "1.0"
-
-[build-dependencies]
-cc = "1.0"
